requirements.txt
Django==4.2.10
djangorestframework
psycopg2-binary
django-cors-headers
django-filter
